+++ cli/cmd/list.go
@@ -2,16 +2,13 @@ package cmd
 
 import (
 	"io"
-	"sort"
 
 	"github.com/chanzuckerberg/happy/cli/pkg/output"
 	stackservice "github.com/chanzuckerberg/happy/cli/pkg/stack_mgr"
-	backend "github.com/chanzuckerberg/happy/shared/backend/aws"
 	"github.com/chanzuckerberg/happy/shared/config"
-	"github.com/chanzuckerberg/happy/shared/diagnostics"
+	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"golang.org/x/exp/maps"
 )
 
 type StructuredListResult struct {
@@ -19,10 +16,13 @@ type StructuredListResult struct {
 	Stacks []stackservice.StackInfo
 }
 
+var listAll bool
+
 func init() {
 	rootCmd.AddCommand(listCmd)
 	config.ConfigureCmdWithBootstrapConfig(listCmd)
 	listCmd.Flags().StringVar(&OutputFormat, "output", "text", "Output format. One of: json, yaml, or text. Defaults to text, which is the only interactive mode.")
+	listCmd.Flags().BoolVar(&listAll, "all", false, "List all stacks, not just those belonging to this app")
 }
 
 var listCmd = &cobra.Command{
@@ -34,57 +34,20 @@ var listCmd = &cobra.Command{
 		if OutputFormat != "text" {
 			logrus.SetOutput(io.Discard)
 		}
-
-		ctx := cmd.Context()
-		bootstrapConfig, err := config.NewBootstrapConfig(cmd)
+		happyClient, err := makeHappyClient(cmd, sliceName, "", []string{}, false)
 		if err != nil {
-			return err
+			return errors.Wrap(err, "unable to initialize the happy client")
 		}
-		happyConfig, err := config.NewHappyConfig(bootstrapConfig)
+
+		stackInfos, err := happyClient.StackService.CollectStackInfo(cmd.Context(), listAll, happyClient.HappyConfig.App())
 		if err != nil {
-			return err
+			return errors.Wrap(err, "unable to collect stack info")
 		}
 
-		b, err := backend.NewAWSBackend(ctx, happyConfig.GetEnvironmentContext())
-		if err != nil {
-			return err
-		}
-
-		workspaceRepo := createWorkspaceRepo(false, b)
-		stackSvc := stackservice.NewStackService().WithHappyConfig(happyConfig).WithBackend(b).WithWorkspaceRepo(workspaceRepo)
-
-		stacks, err := stackSvc.GetStacks(ctx)
-		if err != nil {
-			return err
-		}
-
-		// Iterate in order
-		stackInfos := []stackservice.StackInfo{}
-		stackNames := maps.Keys(stacks)
-		sort.Strings(stackNames)
-		for _, name := range stackNames {
-			stack := stacks[name]
-			stackInfo, err := stack.GetStackInfo(ctx, name)
-			if err != nil {
-				logrus.Warnf("Error retrieving stack %s:  %s", name, err)
-				if !diagnostics.IsInteractiveContext(ctx) {
-					stackInfos = append(stackInfos, stackservice.StackInfo{
-						Name:    name,
-						Status:  "error",
-						Message: err.Error(),
-					})
-				}
-				continue
-			}
-			stackInfos = append(stackInfos, *stackInfo)
-		}
-
-		logrus.Debugf("listing stacks in environment '%s'", happyConfig.GetEnv())
 		printer := output.NewPrinter(OutputFormat)
-
-		err = printer.PrintStacks(ctx, stackInfos)
+		err = printer.PrintStacks(cmd.Context(), stackInfos)
 		if err != nil {
-			return err
+			return errors.Wrap(err, "unable to print stacks")
 		}
 
 		return nil