+++ cli/cmd/update.go
@@ -7,10 +7,10 @@ import (
 	happyCmd "github.com/chanzuckerberg/happy/cli/pkg/cmd"
 	stackservice "github.com/chanzuckerberg/happy/cli/pkg/stack_mgr"
 	"github.com/chanzuckerberg/happy/shared/config"
+	"github.com/chanzuckerberg/happy/shared/options"
 	"github.com/chanzuckerberg/happy/shared/util"
 	"github.com/chanzuckerberg/happy/shared/workspace_repo"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
@@ -44,19 +44,19 @@ var updateCmd = &cobra.Command{
 
 func runUpdate(cmd *cobra.Command, args []string) error {
 	stackName := args[0]
-	happyClient, err := makeHappyClient(cmd, sliceName, stackName, []string{tag}, createTag, dryRun, ModeUpdate)
+	happyClient, err := makeHappyClient(cmd, sliceName, stackName, []string{tag}, createTag)
 	if err != nil {
 		return errors.Wrap(err, "unable to initialize the happy client")
 	}
 
-	ctx := cmd.Context()
+	ctx := context.WithValue(cmd.Context(), options.DryRunKey, dryRun)
 	err = validate(
 		validateConfigurationIntegirty(ctx, happyClient),
 		validateGitTree(happyClient.HappyConfig.GetProjectRoot()),
-		validateTFEBackLog(ctx, dryRun, happyClient.AWSBackend),
+		validateTFEBackLog(ctx, happyClient.AWSBackend),
 		validateStackNameAvailable(ctx, happyClient.StackService, stackName, force),
-		validateStackExistsUpdate(ctx, stackName, dryRun, happyClient),
-		validateECRExists(ctx, stackName, dryRun, terraformECRTargetPathTemplate, happyClient),
+		validateStackExistsUpdate(ctx, stackName, happyClient),
+		validateECRExists(ctx, stackName, terraformECRTargetPathTemplate, happyClient),
 		validateImageExists(ctx, createTag, skipCheckTag, happyClient.ArtifactBuilder),
 	)
 	if err != nil {
@@ -71,13 +71,13 @@ func runUpdate(cmd *cobra.Command, args []string) error {
 	return updateStack(ctx, cmd, stack, force, happyClient)
 }
 
-func validateStackExistsUpdate(ctx context.Context, stackName string, dryRun bool, happyClient *HappyClient) validation {
+func validateStackExistsUpdate(ctx context.Context, stackName string, happyClient *HappyClient) validation {
 	return func() error {
 		// 1.) if the stack does not exist and force flag is used, call the create function first
 		_, err := happyClient.StackService.GetStack(ctx, stackName)
 		if err != nil {
 			if force {
-				_, err = happyClient.StackService.Add(ctx, stackName, dryRun)
+				_, err = happyClient.StackService.Add(ctx, stackName)
 				if err != nil {
 					return errors.Wrap(err, "unable to create the stack")
 				}
@@ -92,7 +92,6 @@ func validateStackExistsUpdate(ctx context.Context, stackName string, dryRun boo
 
 func updateStack(ctx context.Context, cmd *cobra.Command, stack *stackservice.Stack, forceFlag bool, happyClient *HappyClient) error {
 	// 1.) update the workspace's meta variables
-	// TODO: is this used? the only thing I think some old happy environments use is the priority? I guess stack tags too
 	stackMeta, err := updateStackMeta(ctx, stack.Name, happyClient)
 	if err != nil {
 		return errors.Wrap(err, "unable to update the stack's meta information")
@@ -100,24 +99,13 @@ func updateStack(ctx context.Context, cmd *cobra.Command, stack *stackservice.St
 
 	// 2.) apply the terraform for the stack
 	stack = stack.WithMeta(stackMeta)
-	err = stack.Apply(ctx, makeWaitOptions(stack.Name, happyClient.HappyConfig, happyClient.AWSBackend), dryRun, workspace_repo.Message(fmt.Sprintf("Happy %s Update Stack [%s]", util.GetVersion().Version, stack.Name)))
+	err = stack.Apply(ctx, makeWaitOptions(stack.Name, happyClient.HappyConfig, happyClient.AWSBackend), workspace_repo.Message(fmt.Sprintf("Happy %s Update Stack [%s]", util.GetVersion().Version, stack.Name)))
 	if err != nil {
 		return errors.Wrap(err, "failed to apply the stack")
 	}
 
-	if dryRun {
-		if happyClient.Mode == ModeCreate {
-			logrus.Debugf("cleaning up stack '%s'", stack.Name)
-			err = happyClient.StackService.Remove(ctx, stack.Name, false)
-			if err != nil {
-				return errors.Wrap(err, "unable to remove stack")
-			}
-		}
-		return nil
-	}
-
 	// 3.) run migrations tasks
-	shouldRunMigration, err := happyCmd.ShouldRunMigrations(cmd, happyClient.HappyConfig)
+	shouldRunMigration, err := happyCmd.ShouldRunMigrations(ctx, cmd, happyClient.HappyConfig)
 	if err != nil {
 		return err
 	}
@@ -135,23 +123,33 @@ func updateStack(ctx context.Context, cmd *cobra.Command, stack *stackservice.St
 }
 
 func updateStackMeta(ctx context.Context, stackName string, happyClient *HappyClient) (*stackservice.StackMeta, error) {
-	stackMeta := happyClient.StackService.NewStackMeta(stackName)
-	stackMeta.Load(map[string]string{
-		"happy/meta/configsecret": happyClient.HappyConfig.GetSecretId(),
-	})
 	if sliceDefaultTag != "" {
 		happyClient.ArtifactBuilder.WithTags([]string{sliceDefaultTag})
+	}
+	dryRun, ok := ctx.Value(options.DryRunKey).(bool)
+	if !ok {
+		dryRun = false
 	}
 	// for updating and creating (unless in dry-run mode), there should only be one tag (either provided or generated)
 	tag := ""
 	if len(happyClient.ArtifactBuilder.GetTags()) == 1 {
 		tag = happyClient.ArtifactBuilder.GetTags()[0]
-	} else if !happyClient.DryRun {
+	} else if !dryRun {
 		return nil, errors.New("there should only be one tag when updating or creating a stack")
 	}
-	err := stackMeta.Update(ctx, tag, happyClient.StackTags, "", happyClient.StackService)
+	username, err := happyClient.AWSBackend.GetUserName(ctx)
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to update the stack meta")
+		return nil, errors.Wrap(err, "unable to get calling user's name")
 	}
-	return stackMeta, nil
+	stackMeta := happyClient.StackService.NewStackMeta(stackName)
+	return stackMeta.UpdateAll(
+		tag,
+		happyClient.StackTags,
+		"",
+		username,
+		happyClient.HappyConfig.GetProjectRoot(),
+		happyClient.HappyConfig,
+		stackName,
+		happyClient.HappyConfig.GetEnv(),
+	), nil
 }