+++ shared/options/dryrun.go
@@ -1,0 +1,5 @@
+package options
+
+type DryRun string
+
+const DryRunKey DryRun = "dry-run"