+++ shared/workspace_repo/workspace_repo.go
@@ -22,16 +22,16 @@ type Workspace interface {
 	Run(ctx context.Context, options ...TFERunOption) error
 	SetVars(ctx context.Context, key string, value string, description string, sensitive bool) error
 	RunConfigVersion(ctx context.Context, configVersionId string, options ...TFERunOption) error
-	Wait(ctx context.Context, dryRun bool) error
-	WaitWithOptions(ctx context.Context, waitOptions options.WaitOptions, dryRun bool) error
-	ResetCache()
+	Wait(ctx context.Context) error
+	WaitWithOptions(ctx context.Context, waitOptions options.WaitOptions) error
 	GetTags(ctx context.Context) (map[string]string, error)
+	GetHappyMetaRaw(ctx context.Context) ([]byte, error)
 	GetOutputs(ctx context.Context) (map[string]string, error)
 	GetEndpoints(ctx context.Context) (map[string]string, error)
 	GetResources(ctx context.Context) ([]util.ManagedResource, error)
 	GetCurrentRunStatus(ctx context.Context) string
 	GetCurrentRunUrl(ctx context.Context) string
-	UploadVersion(ctx context.Context, targzFilePath string, dryRun bool) (string, error)
+	UploadVersion(ctx context.Context, targzFilePath string) (string, error)
 	SetOutputs(map[string]string)          // For testing purposes only
 	SetClient(tfc *tfe.Client)             // For testing purposes only
 	SetWorkspace(workspace *tfe.Workspace) // For testing purposes only