+++ cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.73.1](https://github.com/chanzuckerberg/happy/compare/cli-v0.73.0...cli-v0.73.1) (2023-04-20)
+
+
+### Miscellaneous Chores
+
+* **cli:** Synchronize happy platform versions
+
+## [0.73.0](https://github.com/chanzuckerberg/happy/compare/cli-v0.72.0...cli-v0.73.0) (2023-04-20)
+
+
+### Features
+
+* filter the stacks by app; display git info ([#1594](https://github.com/chanzuckerberg/happy/issues/1594)) ([665f35c](https://github.com/chanzuckerberg/happy/commit/665f35c39d7eff37ff8d0bca34f37db08f0eb753))
+
 ## [0.72.0](https://github.com/chanzuckerberg/happy/compare/cli-v0.71.0...cli-v0.72.0) (2023-04-20)